worker/requirements.txt
APScheduler==3.10.1
pyodbc==4.0.39
python-dotenv==1.0.0
structlog==23.1.0
numpy==1.24.3
scikit-learn==1.2.2
pandas==2.0.2
aiohttp==3.8.5
aioodbc==0.4.1
python-dateutil==2.8.2
